+++ src/ArduinoJson/JsonPair.hpp
@@ -8,19 +8,45 @@
 
 namespace ARDUINOJSON_NAMESPACE {
 
+class JsonKey {
+ public:
+  JsonKey(Slot* slot) : _slot(slot) {}
+
+  operator const char*() const {
+    return c_str();
+  }
+
+  const char* c_str() const {
+    return _slot ? _slot->key : 0;
+  }
+
+  bool isNull() const {
+    return _slot == 0 || _slot->key == 0;
+  }
+
+  bool isStatic() const {
+    return _slot ? _slot->value.keyIsStatic : true;
+  }
+
+  friend bool operator==(JsonKey lhs, const char* rhs) {
+    if (lhs.isNull()) return rhs == 0;
+    return rhs ? !strcmp(lhs, rhs) : false;
+  }
+
+ private:
+  Slot* _slot;
+};
+
 // A key value pair for JsonObjectData.
 class JsonPair {
  public:
-  JsonPair(MemoryPool* memoryPool, Slot* slot) {
+  JsonPair(MemoryPool* memoryPool, Slot* slot) : _key(slot) {
     if (slot) {
-      _key = slot->key;
       _value = JsonVariant(memoryPool, &slot->value);
-    } else {
-      _key = 0;
     }
   }
 
-  const char* key() const {
+  JsonKey key() const {
     return _key;
   }
 
@@ -29,7 +55,7 @@ class JsonPair {
   }
 
  private:
-  const char* _key;
+  JsonKey _key;
   JsonVariant _value;
 };
 }  // namespace ARDUINOJSON_NAMESPACE