+++ src/ArduinoJson/Strings/FixedSizeFlashString.hpp
@@ -17,12 +17,12 @@ class FixedSizeFlashString {
     return strncmp_P(expected, actual, _size) == 0;
   }
 
-  bool is_null() const {
+  bool isNull() const {
     return !_str;
   }
 
-  template <typename Buffer>
-  const char* save(Buffer* memoryPool) const {
+  template <typename TMemoryPool>
+  const char* save(TMemoryPool* memoryPool) const {
     if (!_str) return NULL;
     void* dup = memoryPool->alloc(_size);
     if (dup != NULL) memcpy_P(dup, (const char*)_str, _size);