+++ src/ArduinoJson/Strings/ZeroTerminatedRamStringConst.hpp
@@ -1,0 +1,43 @@
+// ArduinoJson - arduinojson.org
+// Copyright Benoit Blanchon 2014-2018
+// MIT License
+
+#pragma once
+
+#include <stdlib.h>  // size_t
+#include <string.h>  // strcmp
+
+namespace ARDUINOJSON_NAMESPACE {
+
+class ZeroTerminatedRamStringConst {
+ public:
+  ZeroTerminatedRamStringConst(const char* str) : _str(str) {}
+
+  bool equals(const char* expected) const {
+    const char* actual = _str;
+    if (!actual || !expected) return actual == expected;
+    return strcmp(actual, expected) == 0;
+  }
+
+  bool isNull() const {
+    return !_str;
+  }
+
+  template <typename TMemoryPool>
+  const char* save(TMemoryPool*) const {
+    return _str;
+  }
+
+  size_t size() const {
+    return strlen(_str);
+  }
+
+ protected:
+  const char* _str;
+};
+
+inline ZeroTerminatedRamStringConst makeString(const char* str) {
+  return ZeroTerminatedRamStringConst(str);
+}
+
+}  // namespace ARDUINOJSON_NAMESPACE